+++ test/token/ERC721/extensions/ERC721Votes.test.js
@@ -5,7 +5,7 @@ const { expect } = require('chai');
 
 const { clock, clockFromReceipt } = require('../../../helpers/time');
 
-const { shouldBehaveLikeVotes } = require('../../../governance/utils/Votes.behavior.old');
+const { shouldBehaveLikeVotes } = require('../../../governance/utils/Votes.behavior');
 
 const MODES = {
   blocknumber: artifacts.require('$ERC721Votes'),
